requirements/base.txt
numpy>=1.26
pandas>=2.1
scikit-learn>=1.3
imbalanced-learn>=0.11
pydantic>=2.5
PyYAML>=6.0
joblib>=1.3
python-dotenv>=1.0
